+++ events.py
@@ -28,17 +28,3 @@ class SFDCEventListener(sublime_plugin.EventListener):
         """
 
         util.display_active_project(view)
-
-    def on_modified_async(self, view):
-        """
-        Every time when you modified the context, just hide the console, 
-        you can close it in sublime settings
-        """
-        settings = context.get_settings() 
-
-        # If it is not SFDC Component, just return
-        if not view.file_name(): return
-        name, extension = util.get_file_attr(view.file_name())
-        if extension not in settings["component_extensions"]: return
-
-        if settings["hidden_console_on_modify"]: util.hide_panel()